Channels: move inbound dispatch testkit into contracts
